hmg double, network bug fix

+++ app/src/main/java/telapo/meteorwatcher/dal/model/observation/Cycle.kt
@@ -3,6 +3,6 @@ package telapo.meteorwatcher.dal.model.observation
 data class Cycle(
     val Id: Int,
     val Swarms: MutableList<SwarmResults>,
-    var Hmg: Int,
+    var Hmg: Double,
     var Lm: Int
 )
